requirements.txt
GitPython>=3.1.40
python-dotenv>=1.0.0